% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.Words.R
\name{count.Words}
\alias{count.Words}
\title{Count the most used words of the articles and return a graph of the frequency of the word}
\usage{
count.Words(text)
}
\arguments{
\item{text}{Any tibble with a cloumn named text, you can use the tibble given by the function getText.Article}
}
\value{
The number of time each words is used in the text and a graph
}
\description{
You can use the tibble given by the function getText.Article
}
