% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{The Economist count words}
\usage{
count_words(link)
}
\arguments{
\item{link}{is an element of the vector link or an url of The Economist}
}
\value{
A graph with the most used words, only the words that are used more than 3 times
}
\description{
Count the words of the selected article, use a stopword database imported from tidytext. Reorders the words and displays a graph.
}
\details{
You can use an element of the vector link to chose the articole you want to scrape
}
\examples{
 count_words("united-states/2020/02/11/donald-trump-loosens-americas-restrictions-on-landmines")
}
