% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Get Links of all the articles on The Economist homepage}
\usage{
get_links()
}
\value{
A chr vector with all the links of the article
}
\description{
Get Links of all the articles on The Economist homepage
}
