% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Get the text of an articles chosen among The Economist homepage}
\usage{
get_text(link)
}
\arguments{
\item{link}{is an element of the vector link or an url of The Economist}
}
\value{
A tibble of two variables the first one is the paragraph number and the second the text
}
\description{
You can use an element of the vector link to chose the articole you want to scrape
}
