% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.Articles.R
\name{getLink.Articles}
\alias{getLink.Articles}
\title{Get Links of all the articles on The Economist homepage}
\usage{
getLink.Articles()
}
\arguments{
\item{()}{no param}
}
\value{
A chr vector with all the links of the article
}
\description{
Get Links of all the articles on The Economist homepage
}
