% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{Counts the number of words used in the article and returns a graph}
\usage{
count_words(link)
}
\arguments{
\item{link}{is an element of the vector link or an url of The Economist}
}
\value{
A graph with the most used words, only the words that are used more than 3 times
}
\description{
You can use an element of the vector link to chose the articole you want to scrape
}
